backend/requirements.txt
# backend/requirements.txt

django
djangorestframework
psycopg2-binary
python-dotenv
requests
django-cors-headers